Refactored course-related routes to use new file structure and unified role-based components. Simplified role management by removing `RoleComponentManager` where unnecessary.
