+++ frontend/src/routes/(protected)/courses/$courseId/grades/index.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router'
-import GradesPageStudentView from '@/pages/student/courses/$courseId/grades'
-import CourseOverviewStudentView from '@/pages/student/courses/$courseId'
+import CourseGradesPage from '@/pages/student/courses/$courseId/grades/course-grades.tsx'
+import CourseOverviewStudentView from '@/pages/student/courses/$courseId/course-overview.tsx'
 import RoleComponentManager from '@/components/role-component-manager.tsx'
 import { useAuth } from '@/features/auth/auth.hook.ts'
 
@@ -9,12 +9,5 @@ export const Route = createFileRoute('/(protected)/courses/$courseId/grades/')({
 })
 
 function RouteComponent() {
-  const { authUser } = useAuth('protected')
-
-  return <RoleComponentManager
-    currentRole={authUser.role}
-    roleRender={{
-      student: <GradesPageStudentView />,
-    }}
-  />
+  return <CourseGradesPage />
 }
