+++ frontend/src/routes/(protected)/courses/$courseId/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import CourseOverviewStudentView from '@/pages/student/courses/$courseId'
+import CourseOverview from '@/pages/student/courses/$courseId/course-overview.tsx'
 import RoleComponentManager from '@/components/role-component-manager.tsx'
 import { useAuth } from '@/features/auth/auth.hook.ts'
 
@@ -10,14 +10,7 @@ export const Route = createFileRoute(
 })
 
 function RouteComponent() {
-  const { authUser } = useAuth('protected')
-
   return (
-    <RoleComponentManager
-      currentRole={authUser.role}
-      roleRender={{
-        student: <CourseOverviewStudentView />,
-      }}
-    />
+    <CourseOverview />
   )
 }
