+++ frontend/src/routes/(protected)/courses/$courseId/modules/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import CourseModulesPage from '@/pages/student/courses/$courseId/modules'
+import CourseModulesPage from '@/pages/student/courses/$courseId/modules/course-modules.tsx'
 import { useAuth } from '@/features/auth/auth.hook.ts'
 import RoleComponentManager from '@/components/role-component-manager.tsx'
 
@@ -10,12 +10,5 @@ export const Route = createFileRoute('/(protected)/courses/$courseId/modules/')(
 )
 
 function RouteComponent() {
-  const { authUser } = useAuth('protected')
-
-  return <RoleComponentManager
-    currentRole={authUser.role}
-    roleRender={{
-      student: <CourseModulesPage />,
-    }}
-  />
+  return <CourseModulesPage />
 }
